Remove the old watching infrastructure.
