+++ core/lib/compass/configuration/data.rb
@@ -160,7 +160,7 @@ module Compass
 
       def watch(glob, &block)
         @watches ||= []
-        @watches << Watcher::Watch.new(glob, &block)
+        @watches << Watch.new(glob, &block)
       end
 
       def watches
